[ADD]: Add a new data showing the percentage of SVMs after the LVM is split.

+++ cVMRequest.h
@@ -8,7 +8,7 @@ class cVMRequest
 private:
 	ID request_id;
 	double original_res_request;
-	int  num_svm;
+	uint  num_svm;
 	double svm_res_request;
 	bool splittable;
 	double lambda;
@@ -27,8 +27,8 @@ public:
 	void setOriginalResRequest(double _request){original_res_request = _request;}
 	double getOriginalResRequest(void){return original_res_request;}
 
-	void setSVMNumber(int _num){num_svm = _num;}
-	double getSVMNumber(void){return num_svm;}
+	void setSVMNumber(uint _num){num_svm = _num;}
+	uint getSVMNumber(void){return num_svm;}
 
 	void setSVMResRequest(double _request){svm_res_request = _request;}
 	double getSVMResRequest(void){return svm_res_request;}
@@ -50,7 +50,7 @@ public:
 
 public:
 	cVMRequest(void);
-	cVMRequest(ID _id,double _original_request,double _lambda,uint _arrival_time,uint _duration_time,bool _splittable = false,int _num_svm = 0,int _svm_res_request = 0):\
+	cVMRequest(ID _id,double _original_request,double _lambda,uint _arrival_time,uint _duration_time,bool _splittable = false,uint _num_svm = 0,int _svm_res_request = 0):\
 		request_id(_id),original_res_request(_original_request),\
 		arrival_time(_arrival_time),duration_time(_duration_time),\
 		lambda(_lambda),num_svm(_num_svm),svm_res_request(_svm_res_request),splittable(_splittable){departure_time = arrival_time + duration_time;}
